Extract helper functions and unit tests from IngestRecordsIT
